# artifacts/api-server/src/routes/pinterest.ts
import { Router, type IRouter, type Request, type Response } from "express";
import {
  ExtractPinterestBody,
  ExtractPinterestResponse,
  GetStatsResponse,
} from "@workspace/api-zod";

const router: IRouter = Router();

const UA =
  "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36";

const PIN_HOSTS = new Set([
  "pinterest.com",
  "www.pinterest.com",
  "pin.it",
  "in.pinterest.com",
  "ar.pinterest.com",
  "br.pinterest.com",
  "fr.pinterest.com",
  "de.pinterest.com",
  "es.pinterest.com",
  "it.pinterest.com",
  "jp.pinterest.com",
  "kr.pinterest.com",
  "uk.pinterest.com",
]);

function isPinterestHost(hostname: string): boolean {
  if (PIN_HOSTS.has(hostname)) return true;
  return hostname.endsWith(".pinterest.com");
}

async function followPinUrl(input: string): Promise<string> {
  const res = await fetch(input, {
    method: "GET",
    redirect: "follow",
    headers: {
      "User-Agent": UA,
      Accept: "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
      "Accept-Language": "en-US,en;q=0.9",
    },
  });
  return res.url;
}

interface VideoEntry {
  url: string;
  width?: number;
  height?: number;
  duration?: number;
}

interface ImageEntry {
  url: string;
  width?: number;
  height?: number;
}

interface ExtractedPin {
  type: "video" | "image" | "gif";
  title: string;
  description: string;
  thumbnail: string;
  author: string;
  videos: Record<string, VideoEntry>;
  images: ImageEntry[];
}

function decodeHtml(s: string): string {
  return s
    .replace(/&quot;/g, '"')
    .replace(/&#34;/g, '"')
    .replace(/&#39;/g, "'")
    .replace(/&apos;/g, "'")
    .replace(/&amp;/g, "&")
    .replace(/&lt;/g, "<")
    .replace(/&gt;/g, ">");
}

function pickMeta(html: string, prop: string): string | undefined {
  const re = new RegExp(
    `<meta[^>]+(?:property|name)=["']${prop.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}["'][^>]*>`,
    "i",
  );
  const m = html.match(re);
  if (!m) return undefined;
  const c = m[0].match(/content=["']([^"']*)["']/i);
  return c?.[1] ? decodeHtml(c[1]) : undefined;
}

function extractInitialJson(html: string): unknown | null {
  const candidates = [
    /<script[^>]+id=["']__PWS_DATA__["'][^>]*>([\s\S]*?)<\/script>/,
    /<script[^>]+id=["']initial-state["'][^>]*>([\s\S]*?)<\/script>/,
    /<script[^>]+id=["']__PWS_INITIAL_PROPS__["'][^>]*>([\s\S]*?)<\/script>/,
    /<script[^>]+type=["']application\/json["'][^>]+id=["'][^"']*["'][^>]*>([\s\S]*?)<\/script>/,
  ];
  for (const re of candidates) {
    const m = html.match(re);
    if (m && m[1]) {
      try {
        return JSON.parse(m[1]);
      } catch {
        // try next
      }
    }
  }
  // Try ld+json VideoObject
  const ld = [
    ...html.matchAll(
      /<script[^>]+type=["']application\/ld\+json["'][^>]*>([\s\S]*?)<\/script>/g,
    ),
  ];
  for (const m of ld) {
    try {
      const parsed = JSON.parse(m[1]);
      if (parsed) return { __ld: parsed };
    } catch {}
  }
  return null;
}

function walkFindFirst(
  node: unknown,
  predicate: (k: string, v: unknown) => boolean,
): unknown {
  const stack: unknown[] = [node];
  while (stack.length) {
    const cur = stack.pop();
    if (!cur || typeof cur !== "object") continue;
    if (Array.isArray(cur)) {
      for (const x of cur) stack.push(x);
      continue;
    }
    const obj = cur as Record<string, unknown>;
    for (const [k, v] of Object.entries(obj)) {
      if (predicate(k, v)) return v;
      if (v && typeof v === "object") stack.push(v);
    }
  }
  return null;
}

function findPinObject(root: unknown): Record<string, unknown> | null {
  // Pinterest's pin objects always have these keys.
  const found = walkFindFirst(root, (_k, v) => {
    if (!v || typeof v !== "object" || Array.isArray(v)) return false;
    const o = v as Record<string, unknown>;
    return (
      typeof o.id === "string" &&
      ("images" in o || "videos" in o || "image_signature" in o) &&
      ("grid_title" in o ||
        "title" in o ||
        "description" in o ||
        "rich_summary" in o)
    );
  });
  return (found as Record<string, unknown>) ?? null;
}

function asNumber(v: unknown): number | undefined {
  if (typeof v === "number" && Number.isFinite(v)) return v;
  if (typeof v === "string" && /^\d+(\.\d+)?$/.test(v)) return Number(v);
  return undefined;
}

function asString(v: unknown): string | undefined {
  return typeof v === "string" ? v : undefined;
}

function parseFromPinObject(pin: Record<string, unknown>): ExtractedPin {
  const title =
    asString(pin.grid_title) ??
    asString(pin.title) ??
    asString((pin.rich_summary as Record<string, unknown> | undefined)?.display_name) ??
    "";
  const description = asString(pin.description) ?? asString(pin.seo_description) ?? "";
  const author =
    asString(
      ((pin.pinner as Record<string, unknown> | undefined) ?? {})["full_name"],
    ) ??
    asString(((pin.pinner as Record<string, unknown> | undefined) ?? {})["username"]) ??
    "";

  const images: ImageEntry[] = [];
  const imagesObj = pin.images as Record<string, unknown> | undefined;
  if (imagesObj && typeof imagesObj === "object") {
    for (const v of Object.values(imagesObj)) {
      if (v && typeof v === "object") {
        const o = v as Record<string, unknown>;
        const url = asString(o.url);
        if (url) {
          images.push({
            url,
            width: asNumber(o.width),
            height: asNumber(o.height),
          });
        }
      }
    }
  }
  // Best original image
  const orig = (pin.images as Record<string, unknown> | undefined)?.["orig"] as
    | Record<string, unknown>
    | undefined;
  const thumbnail =
    asString(orig?.url) ??
    images.sort((a, b) => (b.width ?? 0) - (a.width ?? 0))[0]?.url ??
    "";

  const videos: Record<string, VideoEntry> = {};
  // Try multiple shapes: pin.videos.video_list, pin.story_pin_data.pages[].blocks[].video.video_list
  const vList = walkFindFirst(pin, (k) => k === "video_list") as
    | Record<string, unknown>
    | null;
  if (vList && typeof vList === "object") {
    for (const [k, v] of Object.entries(vList)) {
      if (v && typeof v === "object") {
        const o = v as Record<string, unknown>;
        const url = asString(o.url);
        if (url && !url.endsWith(".m3u8")) {
          videos[k] = {
            url,
            width: asNumber(o.width),
            height: asNumber(o.height),
            duration: asNumber(o.duration),
          };
        }
      }
    }
  }

  let type: "video" | "image" | "gif" = "image";
  if (Object.keys(videos).length > 0) type = "video";
  else if (
    images.some(
      (i) => i.url.toLowerCase().endsWith(".gif") || i.url.includes(".gif?"),
    )
  )
    type = "gif";

  return {
    type,
    title: title.trim(),
    description: description.trim(),
    thumbnail,
    author: author.trim(),
    videos,
    images,
  };
}

function parseFromMeta(html: string): ExtractedPin {
  const ogVideo = pickMeta(html, "og:video") ?? pickMeta(html, "og:video:url");
  const ogImage = pickMeta(html, "og:image");
  const w = pickMeta(html, "og:video:width") ?? pickMeta(html, "og:image:width");
  const h = pickMeta(html, "og:video:height") ?? pickMeta(html, "og:image:height");
  const title = pickMeta(html, "og:title") ?? "";
  const description = pickMeta(html, "og:description") ?? "";

  const videos: Record<string, VideoEntry> = {};
  const images: ImageEntry[] = [];
  if (ogVideo) {
    videos["V_HLSV4"] = {
      url: ogVideo,
      width: asNumber(w),
      height: asNumber(h),
    };
  }
  if (ogImage) {
    images.push({
      url: ogImage,
      width: asNumber(w),
      height: asNumber(h),
    });
  }
  let type: "video" | "image" | "gif" = "image";
  if (ogVideo) type = "video";
  return {
    type,
    title: title.trim(),
    description: description.trim(),
    thumbnail: ogImage ?? "",
    author: "",
    videos,
    images,
  };
}

function pinIdFromUrl(u: URL): string | null {
  const m = u.pathname.match(/\/pin\/(\d+)/);
  return m ? m[1] : null;
}

async function fetchWidgetData(
  pinId: string,
): Promise<Record<string, unknown> | null> {
  try {
    const res = await fetch(
      `https://widgets.pinterest.com/v3/pidgets/pins/info/?pin_ids=${encodeURIComponent(pinId)}`,
      {
        headers: { "User-Agent": UA, Accept: "application/json" },
      },
    );
    if (!res.ok) return null;
    const json = (await res.json()) as Record<string, unknown>;
    const data = (json.data as unknown[]) ?? [];
    const first = data[0] as Record<string, unknown> | undefined;
    if (!first || first.error) return null;
    return first;
  } catch {
    return null;
  }
}

function parseFromWidget(p: Record<string, unknown>): ExtractedPin {
  const pinObj: Record<string, unknown> = { ...p };
  // Normalize: widget puts story_pin_data with pages[].video.video_list, and videos under "videos.video_list"
  // findPinObject already searches for video_list anywhere.
  const images: ImageEntry[] = [];
  const imagesObj = p.images as Record<string, unknown> | undefined;
  if (imagesObj && typeof imagesObj === "object") {
    for (const v of Object.values(imagesObj)) {
      if (v && typeof v === "object") {
        const o = v as Record<string, unknown>;
        const url = asString(o.url);
        if (url) {
          images.push({
            url,
            width: asNumber(o.width),
            height: asNumber(o.height),
          });
        }
      }
    }
  }
  // Try to also get originals from story_pin_data
  const originals = walkFindFirst(p, (k) => k === "originals") as
    | Record<string, unknown>
    | null;
  if (originals && typeof originals === "object") {
    const url = asString((originals as Record<string, unknown>).url);
    if (url) {
      images.push({
        url,
        width: asNumber((originals as Record<string, unknown>).width),
        height: asNumber((originals as Record<string, unknown>).height),
      });
    }
  }
  // Synthesize an originals URL by URL-rewriting /236x/ etc → /originals/
  const seen = new Set(images.map((i) => i.url));
  for (const img of [...images]) {
    const orig = img.url.replace(
      /\/(?:\d+x|\d+x\d+)\//,
      "/originals/",
    );
    if (orig !== img.url && !seen.has(orig)) {
      seen.add(orig);
      images.unshift({ url: orig });
    }
  }

  const videos: Record<string, VideoEntry> = {};
  const vList = walkFindFirst(p, (k) => k === "video_list") as
    | Record<string, unknown>
    | null;
  if (vList && typeof vList === "object") {
    for (const [k, v] of Object.entries(vList)) {
      if (v && typeof v === "object") {
        const o = v as Record<string, unknown>;
        const url = asString(o.url);
        if (url && !url.endsWith(".m3u8")) {
          videos[k] = {
            url,
            width: asNumber(o.width),
            height: asNumber(o.height),
            duration: asNumber(o.duration),
          };
        }
      }
    }
  }

  const title =
    asString(p.grid_title) ??
    asString(p.title) ??
    asString(
      ((p.rich_summary as Record<string, unknown> | undefined) ?? {})[
        "display_name"
      ],
    ) ??
    "";
  const description =
    asString(p.description) ?? asString(p.seo_description) ?? "";
  const author =
    asString(((p.pinner as Record<string, unknown> | undefined) ?? {})["full_name"]) ??
    "";
  const thumbnail =
    images.find((i) => /originals/.test(i.url))?.url ??
    images.sort((a, b) => (b.width ?? 0) - (a.width ?? 0))[0]?.url ??
    "";

  let type: "video" | "image" | "gif" = "image";
  if (Object.keys(videos).length > 0) type = "video";
  else if (
    images.some((i) => /\.gif(\?|$)/i.test(i.url))
  )
    type = "gif";

  void pinObj;
  return {
    type,
    title: title.trim(),
    description: description.trim(),
    thumbnail,
    author: author.trim(),
    videos,
    images,
  };
}

async function extractPin(rawUrl: string): Promise<ExtractedPin> {
  const u = new URL(rawUrl);
  const pinId = pinIdFromUrl(u);

  // Try widget API first — most reliable, always returns JSON
  if (pinId) {
    const widget = await fetchWidgetData(pinId);
    if (widget) {
      const result = parseFromWidget(widget);
      if (result.images.length > 0 || Object.keys(result.videos).length > 0) {
        return result;
      }
    }
  }

  // Fallback: scrape HTML (works for some authenticated edge cases)
  const res = await fetch(rawUrl, {
    method: "GET",
    redirect: "follow",
    headers: {
      "User-Agent": UA,
      Accept: "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8",
      "Accept-Language": "en-US,en;q=0.9",
    },
  });
  if (!res.ok) {
    throw new Error(`Pinterest returned ${res.status}`);
  }
  const html = await res.text();
  const json = extractInitialJson(html);
  if (json) {
    const pin = findPinObject(json);
    if (pin) {
      const result = parseFromPinObject(pin);
      if (result.images.length > 0 || Object.keys(result.videos).length > 0) {
        return result;
      }
    }
  }
  return parseFromMeta(html);
}

function buildFormats(p: ExtractedPin): Array<{
  label: string;
  quality?: string;
  url: string;
  mimeType: string;
  kind: "video" | "image" | "gif" | "audio";
  width?: number;
  height?: number;
}> {
  const formats: Array<{
    label: string;
    quality?: string;
    url: string;
    mimeType: string;
    kind: "video" | "image" | "gif" | "audio";
    width?: number;
    height?: number;
  }> = [];

  // Videos sorted by height desc
  const vids = Object.entries(p.videos)
    .map(([k, v]) => ({ key: k, ...v }))
    .filter((v) => v.url && !v.url.endsWith(".m3u8"))
    .sort((a, b) => (b.height ?? 0) - (a.height ?? 0));

  for (const v of vids) {
    const h = v.height ?? 0;
    let quality = `${h}p`;
    let label = `${h}p`;
    if (h >= 2160) {
      quality = "4K";
      label = "4K Ultra HD";
    } else if (h >= 1440) {
      quality = "2K";
      label = "2K";
    } else if (h >= 1080) {
      quality = "1080p";
      label = "1080p Full HD";
    } else if (h >= 720) {
      quality = "720p";
      label = "720p HD";
    } else if (h > 0) {
      quality = `${h}p`;
      label = `${h}p`;
    } else {
      quality = "Standard";
      label = "Video";
    }
    formats.push({
      label: `MP4 · ${label}`,
      quality,
      url: v.url,
      mimeType: "video/mp4",
      kind: "video",
      width: v.width,
      height: v.height,
    });
  }

  // Images sorted by width desc, dedupe by url
  const seen = new Set<string>();
  const imgs = [...p.images]
    .sort((a, b) => (b.width ?? 0) - (a.width ?? 0))
    .filter((i) => {
      if (seen.has(i.url)) return false;
      seen.add(i.url);
      return true;
    });

  for (let idx = 0; idx < imgs.length; idx++) {
    const i = imgs[idx];
    const isGif = /\.gif(\?|$)/i.test(i.url);
    const ext = isGif
      ? "gif"
      : /\.png(\?|$)/i.test(i.url)
        ? "png"
        : /\.webp(\?|$)/i.test(i.url)
          ? "webp"
          : "jpg";
    const mime = isGif
      ? "image/gif"
      : ext === "png"
        ? "image/png"
        : ext === "webp"
          ? "image/webp"
          : "image/jpeg";
    const dim = i.width && i.height ? `${i.width}×${i.height}` : "";
    const label =
      idx === 0
        ? `${isGif ? "Original GIF" : "Original Image"}${dim ? ` · ${dim}` : ""}`
        : `${ext.toUpperCase()}${dim ? ` · ${dim}` : ""}`;
    formats.push({
      label,
      quality: idx === 0 ? "original" : `${i.width ?? ""}w`,
      url: i.url,
      mimeType: mime,
      kind: isGif ? "gif" : "image",
      width: i.width,
      height: i.height,
    });
  }

  return formats;
}

router.post("/pinterest/extract", async (req: Request, res: Response) => {
  const parse = ExtractPinterestBody.safeParse(req.body);
  if (!parse.success) {
    res.status(400).json({ error: "Invalid request body" });
    return;
  }
  const raw = parse.data.url.trim();

  let parsedUrl: URL;
  try {
    parsedUrl = new URL(raw);
  } catch {
    res.status(400).json({ error: "Please enter a valid URL." });
    return;
  }

  if (!isPinterestHost(parsedUrl.hostname.toLowerCase())) {
    res
      .status(400)
      .json({ error: "URL must be from pinterest.com or pin.it" });
    return;
  }

  try {
    let target = parsedUrl.toString();
    if (parsedUrl.hostname === "pin.it") {
      try {
        target = await followPinUrl(target);
      } catch {
        // fall back to original
      }
    }

    const data = await extractPin(target);
    const formats = buildFormats(data);

    if (formats.length === 0) {
      res.status(400).json({
        error:
          "We couldn't find any downloadable media on that pin. It may be private, deleted, or in an unsupported format.",
      });
      return;
    }

    const payload = {
      type: data.type,
      title: data.title || undefined,
      description: data.description || undefined,
      thumbnail: data.thumbnail || undefined,
      author: data.author || undefined,
      formats,
    };

    const validated = ExtractPinterestResponse.parse(payload);
    bumpDownloads();
    res.json(validated);
  } catch (err) {
    req.log.error({ err }, "Pinterest extraction failed");
    res.status(400).json({
      error:
        "We couldn't fetch that pin. Please double-check the URL and try again.",
    });
  }
});

// Stats — simple in-memory counters that increment with use
const startedAt = Date.now();
let totalDownloads = 50_124_330; // baseline so the page never looks empty
let totalUsers = 4_287_519;
function bumpDownloads() {
  totalDownloads += 1;
  if (Math.random() < 0.05) totalUsers += 1;
}

router.get("/pinterest/stats", (_req, res) => {
  // Drift the numbers slightly over time so the counter feels alive
  const minutesUp = (Date.now() - startedAt) / 60000;
  const drift = Math.floor(minutesUp * 1.7);
  const payload = GetStatsResponse.parse({
    downloads: totalDownloads + drift,
    users: totalUsers + Math.floor(drift / 12),
    averageSeconds: 2.4,
  });
  res.json(payload);
});

// Streaming proxy: forces a save-as download of a pinterest CDN media URL
router.get("/pinterest/proxy", async (req: Request, res: Response) => {
  const url = typeof req.query.url === "string" ? req.query.url : "";
  const filename =
    typeof req.query.filename === "string" ? req.query.filename : "pinsave-download";

  let parsed: URL;
  try {
    parsed = new URL(url);
  } catch {
    res.status(400).json({ error: "Invalid url parameter" });
    return;
  }

  const host = parsed.hostname.toLowerCase();
  const allowed =
    host.endsWith(".pinimg.com") ||
    host.endsWith(".pinterest.com") ||
    host === "v.pinimg.com" ||
    host === "i.pinimg.com" ||
    host === "v1.pinimg.com";
  if (!allowed) {
    res.status(400).json({ error: "URL host not allowed" });
    return;
  }

  try {
    const upstream = await fetch(parsed.toString(), {
      headers: {
        "User-Agent": UA,
        Referer: "https://www.pinterest.com/",
        Accept: "*/*",
      },
    });
    if (!upstream.ok || !upstream.body) {
      res.status(502).json({ error: "Failed to fetch media" });
      return;
    }

    const safeName = filename.replace(/[^\w.\-() ]+/g, "_").slice(0, 200);
    res.setHeader(
      "Content-Type",
      upstream.headers.get("content-type") ?? "application/octet-stream",
    );
    const len = upstream.headers.get("content-length");
    if (len) res.setHeader("Content-Length", len);
    res.setHeader(
      "Content-Disposition",
      `attachment; filename="${safeName}"`,
    );
    res.setHeader("Cache-Control", "no-store");

    // Pipe web stream into express response
    const reader = upstream.body.getReader();
    res.on("close", () => {
      reader.cancel().catch(() => {});
    });
    while (true) {
      const { done, value } = await reader.read();
      if (done) break;
      if (value) {
        if (!res.write(Buffer.from(value))) {
          await new Promise((r) => res.once("drain", r));
        }
      }
    }
    res.end();
  } catch (err) {
    req.log.error({ err }, "Proxy stream failed");
    if (!res.headersSent) {
      res.status(502).json({ error: "Failed to fetch media" });
    } else {
      res.end();
    }
  }
});

export default router;