# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: pinterest
    description: Pinterest media extraction
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /pinterest/extract:
    post:
      operationId: extractPinterest
      tags: [pinterest]
      summary: Extract downloadable media from a Pinterest pin URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractRequest"
      responses:
        "200":
          description: Extracted media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractResponse"
        "400":
          description: Invalid URL or extraction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pinterest/stats:
    get:
      operationId: getStats
      tags: [pinterest]
      summary: Service statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ExtractRequest:
      type: object
      properties:
        url:
          type: string
          description: A Pinterest pin URL (pin.it short link or pinterest.com/pin/...)
      required: [url]
    MediaFormat:
      type: object
      properties:
        label:
          type: string
          description: Human-readable label, e.g. "1080p HD", "Original Image", "MP3"
        quality:
          type: string
          description: Quality indicator (e.g., "1080p", "720p", "original")
        url:
          type: string
          description: Direct media URL (use the proxy download endpoint to save with proper filename)
        mimeType:
          type: string
        kind:
          type: string
          enum: [video, image, gif, audio]
        width:
          type: integer
        height:
          type: integer
      required: [label, url, mimeType, kind]
    ExtractResponse:
      type: object
      properties:
        type:
          type: string
          enum: [video, image, gif]
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        author:
          type: string
        formats:
          type: array
          items:
            $ref: "#/components/schemas/MediaFormat"
      required: [type, formats]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
    StatsResponse:
      type: object
      properties:
        downloads:
          type: integer
        users:
          type: integer
        averageSeconds:
          type: number
      required: [downloads, users, averageSeconds]