# lib/api-zod/src/generated/types/extractRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractRequest {
  /** A Pinterest pin URL (pin.it short link or pinterest.com/pin/...) */
  url: string;
}