# lib/api-zod/src/generated/types/extractResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractResponseType } from "./extractResponseType";
import type { MediaFormat } from "./mediaFormat";

export interface ExtractResponse {
  type: ExtractResponseType;
  title?: string;
  description?: string;
  thumbnail?: string;
  author?: string;
  formats: MediaFormat[];
}