# lib/api-zod/src/generated/types/mediaFormat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaFormatKind } from "./mediaFormatKind";

export interface MediaFormat {
  /** Human-readable label, e.g. "1080p HD", "Original Image", "MP3" */
  label: string;
  /** Quality indicator (e.g., "1080p", "720p", "original") */
  quality?: string;
  /** Direct media URL (use the proxy download endpoint to save with proper filename) */
  url: string;
  mimeType: string;
  kind: MediaFormatKind;
  width?: number;
  height?: number;
}